% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acca}
\alias{acca}
\alias{acca.cmatrix}
\alias{acca.matrix}
\title{Average correlation clustering algorithm}
\usage{
acca(m, ...)

\method{acca}{cmatrix}(m, k, maxrep = 2L, maxiter = 100L)

\method{acca}{matrix}(m, k, maxrep = 2L, maxiter = 100L)
}
\arguments{
\item{m}{[\code{matrix(1)}]\cr correlation matrix from \code{\link{corr_matrix}}}

\item{k}{[\code{integer(1)}]\cr number of clusters considered.}

\item{maxrep}{[\code{integer(1)}]\cr maximum number of interactions without change in the clusters.}

\item{maxiter}{[\code{integer(1)}]\cr maximum number of interactions.}
}
\description{
A C++ implementation of the ACCA method
that works directly with the correlation matrix derived from the \code{\link{corr_matrix}} function.
In this sense, this implementation differs from the original,
it works with mixed data and several correlation methods
}
\references{
Bhattacharya, Anindya, and Rajat K. De.
"Average correlation clustering algorithm (ACCA) for grouping of co-regulated
genes with similar pattern of variation in their expression values."
Journal of Biomedical Informatics 43.4 (2010): 560-568.
}
\author{
Igor D.S. Siciliani
}
\keyword{,}
\keyword{acca}
\keyword{correlation}
