% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_required_argument}
\alias{assert_required_argument}
\title{Assert Required Argument}
\usage{
assert_required_argument(arg, description)
}
\arguments{
\item{arg}{[\code{any}]\cr
The argument to check.}

\item{description}{[\code{character(1)}]\cr
A description of the argument's purpose and requirements.}
}
\value{
Throws an error if the argument is missing; otherwise, returns \code{NULL}.
}
\description{
Ensures that a required argument is provided. If the argument is missing, it throws an error with a clear message.
}
