% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_acca.R
\name{best_acca}
\alias{best_acca}
\alias{best_acca.cmatrix}
\alias{best_acca.matrix}
\title{Silhouette (clustering)}
\usage{
best_acca(m, ...)

\method{best_acca}{cmatrix}(m, mink, maxk, maxrep = 2L, maxiter = 100L, ...)

\method{best_acca}{matrix}(m, mink, maxk, maxrep = 2L, maxiter = 100L, ...)
}
\arguments{
\item{m}{[\code{matrix}]\cr correlation matrix
from \code{\link{corr_matrix}}.}

\item{...}{Not used. Included for S3 method consistency.}

\item{mink}{[\code{integer(1)}]\cr minimum number of clusters considered.}

\item{maxk}{[\code{integer(1)}]\cr maximum number of clusters considered.}

\item{maxrep}{[\code{integer(1)}]\cr
Maximum number of consecutive iterations without changes in the cluster structure before stopping the algorithm.
If the clusters remain identical for \code{maxrep} iterations in a row, the algorithm considers that it has converged and stops early, without reaching \code{maxiter}.
It serves to speed up execution by detecting when the clustering has stabilized.}

\item{maxiter}{[\code{integer(1)}]\cr
Maximum number of iterations allowed for the algorithm.
Regardless of cluster stability, the algorithm will be forcibly stopped after \code{maxiter} iterations to prevent infinite loops or excessive computation.
It acts as a hard upper limit on the runtime.}
}
\value{
[\code{list}]\cr
A \code{list} which is a list that has three elements:
\itemize{
\item \strong{silhouette.ave}: The silhouettes average per k.
\item \strong{k}: The sequence of clusters tested.
\item \strong{best.k}: The optimal number of clusters.
}
}
\description{
Determining the optimal number of
cluster in the ACCA clustering using the
average silhouette approach.
}
\examples{

x <- corrp::corrp(iris)
m <- corrp::corr_matrix(x)
best_acca(m, 2, 6)

}
\references{
Leonard Kaufman; Peter J. Rousseeuw (1990).
Finding groups in data : An introduction to cluster analysis.
Hoboken, NJ: Wiley-Interscience.
p. 87. doi:10.1002/9780470316801. ISBN 9780471878766.

Starczewski, Artur, and Adam Krzyżak.
"Performance evaluation of the silhouette index.
"International Conference on Artificial Intelligence
and Soft Computing. Springer, Cham, 2015.
}
\seealso{
\code{\link{sil_acca}}
}
\author{
Igor D.S. Siciliani, Paulo H. dos Santos
}
\keyword{acca,}
\keyword{k}
\keyword{optimal,}
\keyword{silhouette,}
