% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_fun.R
\name{cor_fun}
\alias{cor_fun}
\title{Compute Correlation type analysis with Statistical Significance.}
\usage{
cor_fun(
  df,
  nx,
  ny,
  p.value,
  verbose = TRUE,
  ptest.n.sum = 1000,
  ptest.r = F,
  comp = c("greater", "less"),
  alternative = c("two.sided", "less", "greater"),
  cor.nn = c("pearson", "mic", "dcor", "pps"),
  cor.nc = c("lm", "pps"),
  cor.cc = c("cramersV", "uncoef", "pps"),
  lm.args = list(),
  pearson.args = list(),
  dcor.args = list(),
  mic.args = list(),
  pps.args = list(),
  cramersV.args = list(),
  uncoef.args = list(),
  ...
)
}
\arguments{
\item{df}{input data frame.}

\item{p.value}{\[\code{logical(1)}\]\cr
P-value probability of obtaining the observed results of a test,
assuming that the null hypothesis is correct. By default p.value=0.05 (Cutoff value for p-value.).}

\item{verbose}{\[\code{logical(1)}\]\cr Activate verbose mode.}

\item{comp}{\[\code{character(1)}\]\cr The param \code{p.value} must be greater
or less than those estimated in tests and correlations.}

\item{alternative}{\[\code{character(1)}\]\cr a character string specifying the alternative hypothesis for
the correlation inference. It must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{...}{Additional arguments (TODO).}
}
\description{
Compute one correlation type analysis on two mixed classes columns of a given dataframe.
  The dataframe is allowed to have columns of these four classes: integer,
  numeric, factor and character. The character column is considered as
  categorical variable. This method is original based on Srikanth KS (talegari) cor2 function.
}
\examples{
\dontrun{
air_cor = corrp(airquality)
corrplot::corrplot(air_cor)
corrgram::corrgram(air_cor)
}

}
\references{
KS Srikanth,sidekicks,cor2, 2020.
URL \url{https://github.com/talegari/sidekicks/}.
}
\author{
IDS siciliani (igor-siciliani)
}
\keyword{(http://choosealicense.com/licenses/agpl-3.0/)}
\keyword{AGPLv3}
\keyword{GNU}
