% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\alias{corr_matrix.default}
\alias{corr_matrix.clist}
\title{Create Correlation Matrix from corrp inferences}
\usage{
corr_matrix(c, ...)

\method{corr_matrix}{default}(c, col = c("infer.value", "stat.value", "isig"), isig = TRUE, ...)

\method{corr_matrix}{clist}(c, col = c("infer.value", "stat.value", "isig"), isig = TRUE, ...)
}
\arguments{
\item{c}{\[\code{corrp.list(1)}\]\cr output from the \code{\link{corrp}} function.}

\item{...}{Additional arguments (TODO).}

\item{col}{\[\code{character(1)}\]\cr choose the column to be used in the correlation matrix.}

\item{isig}{\[\code{logical(1)}\]\cr values that are not statistically significant will
be represented by NA or FALSE in the correlation matrix.}
}
\description{
Through the results obtained from corrp function
create a correlation matrix.
}
\examples{
\dontrun{

air_cor = corrp(airquality)
air_m = corr_matrix(air_cor,isig = F)
corrplot::corrplot(air_m)

}

}
\author{
Igor D.S. Siciliani
}
\keyword{,}
\keyword{correlation}
\keyword{corrp}
\keyword{matrix}
