% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_rm.R
\name{corr_rm}
\alias{corr_rm}
\alias{corr_rm.clist}
\alias{corr_rm.list}
\alias{corr_rm.cmatrix}
\alias{corr_rm.matrix}
\title{Filter Data Frame by Correlation Values}
\usage{
corr_rm(df, c, ...)

\method{corr_rm}{clist}(
  df,
  c,
  col = c("infer.value", "stat.value"),
  isig = TRUE,
  cutoff = 0.75,
  ...
)

\method{corr_rm}{list}(
  df,
  c,
  col = c("infer.value", "stat.value"),
  isig = TRUE,
  cutoff = 0.75,
  ...
)

\method{corr_rm}{cmatrix}(df, c, cutoff = 0.75, ...)

\method{corr_rm}{matrix}(df, c, cutoff = 0.75, ...)
}
\arguments{
\item{df}{[\code{data.frame(1)}]\cr The input data frame.}

\item{c}{[\code{clist(1)} | \code{cmatrix}]\cr A correlation list output from the \code{\link[corrp]{corrp}} function (with class \code{clist}),
or a correlation matrix output from the \code{\link[corrp]{corr_matrix}} function (with class \code{cmatrix}).}

\item{...}{Additional arguments.}

\item{col}{[\code{character(1)}]\cr The column to be used in the correlation matrix.}

\item{isig}{[\code{logical(1)}]\cr Whether values that are not statistically significant should
be represented by \code{NA} or \code{FALSE} in the correlation matrix.}

\item{cutoff}{[\code{numeric(1)}]\cr A numeric value for the pairwise absolute correlation cutoff.
The default value is 0.75.}
}
\description{
Remove highly correlated variables from a data frame using the
correlation functions' outputs and the \code{\link[caret]{findCorrelation}} function from the caret package.
}
\examples{

iris_clist <- corrp(iris)
iris_cmatrix <- corr_matrix(iris_clist)
corr_rm(df = iris, c = iris_clist, cutoff = 0.75, col = "infer.value", isig = FALSE)
corr_rm(df = iris, c = iris_cmatrix, cutoff = 0.75, col = "infer.value", isig = FALSE)

}
\author{
Igor D.S. Siciliani, Paulo H. dos Santos
}
\keyword{clist}
\keyword{cmatrix,}
\keyword{correlated,}
\keyword{highly}
