% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{dcor_t_test}
\alias{dcor_t_test}
\title{Create Correlation Matrix from corrp inferences}
\usage{
dcor_t_test(x, y)
}
\arguments{
\item{x}{[\code{data.frame | matrix}]\cr Data of the first sample.}

\item{y}{[\code{data.frame | matrix}]\cr Data of the second sample.}
}
\value{
[\code{list}]\cr returns a list containing:
\item{method}{description of test.}
\item{statistic}{observed value of the test statistic.}
\item{parameter}{degrees of freedom.}
\item{estimate}{bias corrected squared measure of distance correlation between x and y.}
\item{p.value}{p-value of the t-test.}
\item{data.name}{description of data.}
}
\description{
Performs a distance correlation t-test of multivariate independence, based on the distance correlation t-statistic introduced by Székely and Rizzo (2013). This implementation executes the statistical approach of \code{energy::dcorT.test}, but is written in C++ for improved performance. The function supports testing for dependence between multivariate random vectors and uses a t-distribution approximation under the null hypothesis of independence. Empirical benchmarks indicate that this C++ version can be significantly faster than the original R version in the \code{energy} package, especially for larger sample sizes (e.g., 1.5–3x faster for n = 10,000).
}
\examples{
dcor_t_test(iris[, "Petal.Length"], iris[, c("Sepal.Length", "Sepal.Width")])

}
\references{
Székely, G. J., & Rizzo, M. L. (2013). The distance correlation t-test of independence in high dimension. \emph{Journal of Multivariate Analysis}, 117, 193–213. URL: \url{https://doi.org/10.1016/j.jmva.2013.02.012}.
Székely, G. J., & Rizzo, M. L. (2014). Energy statistics: A class of statistics based on distances. \emph{Journal of Statistical Planning and Inference}, 143(8), 1249–1272. URL: \url{https://doi.org/10.1016/j.jspi.2013.03.018}.
}
