% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{dcorT_test}
\alias{dcorT_test}
\title{Create Correlation Matrix from corrp inferences}
\usage{
dcorT_test(x, y)
}
\arguments{
\item{x}{\[\code{data.frame(1) | matrix(1)}]\cr A data of the first sample.}

\item{y}{\[\code{data.frame(1) | matrix(1)}]\cr A data of the second sample.}
}
\value{
returns a list containing
  \item{method}{description of test}
  \item{statistic}{observed value of the test statistic}
  \item{parameter}{degrees of freedom}
  \item{estimate}{(bias corrected) squared dCor(x,y)}
  \item{p.value}{p-value of the t-test}
  \item{data.name}{description of data}
}
\description{
Distance correlation t-test of multivariate independence for high dimension. C++ version of energy::dcorT.test.
}
