% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{ptest}
\alias{ptest}
\title{simple pertumation test}
\usage{
ptest(
  x,
  y,
  FUN,
  alternative = c("two.sided", "less", "greater"),
  num.s = 1000,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{num.s}{number of samples with replacement created with y numeric vector.}
}
\description{
Execute one-sample simple permutation test on two numeric vector.
Its keep one vector constant and ‘shuffle’ the other by resampling.
This approximates the null hypothesis — that there is no dependency between the variables.
}
