% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_corrP.R
\name{rh_corrP}
\alias{rh_corrP}
\title{rh_corrP}
\usage{
rh_corrP(df, corrmat, cutoff = 0.75)
}
\arguments{
\item{df}{input data frame}

\item{corrmat}{correlation matrix from corrP function}

\item{cutoff}{A numeric value for the pair-wise absolute correlation cutoff.
The default values is 0.75.}
}
\description{
remove highly correlated variables from a data.frame using the
corrp correlation matrix and the caret package function `caret::findCorrelation` .
}
\examples{
\dontrun{
air_cor = corrP(airquality)
airqualityH = rh_corrP(df=airquality,corrmat=air_cor,cutoff=0.75)
}
}
\author{
IDS siciliani (igor-siciliani)
}
\keyword{(http://choosealicense.com/licenses/agpl-3.0/)}
\keyword{AGPLv3}
\keyword{GNU}
