% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_corr.R
\name{rm_corr}
\alias{rm_corr}
\alias{rm_corr.clist}
\alias{rm_corr.list}
\alias{rm_corr.cmatrix}
\alias{rm_corr.matrix}
\title{rh_corrP}
\usage{
rm_corr(df, corr, ...)

\method{rm_corr}{clist}(
  df,
  corr,
  col = c("infer.value", "stat.value", "isig"),
  isig = TRUE,
  cutoff = 0.75,
  ...
)

\method{rm_corr}{list}(
  df,
  corr,
  col = c("infer.value", "stat.value", "isig"),
  isig = TRUE,
  cutoff = 0.75,
  ...
)

\method{rm_corr}{cmatrix}(df, corr, cutoff = 0.75, ...)

\method{rm_corr}{matrix}(df, corr, cutoff = 0.75, ...)
}
\arguments{
\item{df}{\[\code{data.frame(1)}\]\cr input data frame}

\item{corr}{\[\code{clist(1)} | \code{cmatrix(1)}\]\cr correlation list from corrp function or matrix from corr_matrix function.}

\item{...}{others parameters.}

\item{cutoff}{\[\code{numeric(1)}\]\cr A numeric value for the pair-wise absolute correlation cutoff.
The default values is 0.75.}
}
\description{
remove highly correlated variables from a data.frame using the
corrp correlation matrix and the caret package function `caret::findCorrelation` .
}
\examples{
\dontrun{
air_cor = corrp(airquality)
airqualityH = rm_corr(df=airquality,corr=air_cor,cutoff=0.75,col = 'infer.value',isig = F)
}
}
\author{
IDS siciliani (igor-siciliani)
}
\keyword{,}
\keyword{clist}
\keyword{cmatrix}
\keyword{correlated,}
\keyword{highly}
