% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_arguments}
\alias{set_arguments}
\title{Set Argument}
\usage{
set_arguments(args_list)
}
\arguments{
\item{args_list}{[\code{list}]\cr
A named list of arguments to be assigned to the parent environment.}
}
\value{
A modified \code{args_list} with the arguments that were assigned to the parent environment removed.
}
\description{
Assigns provided arguments from the \code{args_list} to the parent environment. If an argument is inside the arguments of the methods that calculate statistics, it assigns it on the parent environment, and removes the argument from the list.
}
