% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sil_acca.R
\name{sil_acca}
\alias{sil_acca}
\alias{sil_acca.acca_list}
\alias{sil_acca.list}
\title{Silhouette (Clustering)}
\usage{
sil_acca(acca, m, ...)

\method{sil_acca}{acca_list}(acca, m, ...)

\method{sil_acca}{list}(acca, m, ...)
}
\arguments{
\item{acca}{[\code{acca_list(1)}]\cr Acca clustering results from \code{\link{acca}}.}

\item{m}{[\code{cmatrix(1)|matrix(1)}]\cr A correlation matrix from \code{\link{corr_matrix}}.
By default, the distance matrix (dist) used in this method is given by \code{dist = 1 - m}.}

\item{...}{Additional arguments.}
}
\value{
[\code{numeric(1)}]\cr The average value of
the silhouette width across all data in the entire dataset.
Observations with a large average silhouette width (close to 1)
are very well clustered.
}
\description{
A C++ implementation of the Silhouette method
for interpreting and validating consistency within acca clusters of data.
}
\examples{

x <- corrp::corrp(iris)
m <- corrp::corr_matrix(x)
acca <- corrp::acca(m, 2)
sil_acca(acca, m)

}
\references{
Leonard Kaufman; Peter J. Rousseeuw (1990).
Finding Groups in Data: An Introduction to Cluster Analysis.
Hoboken, NJ: Wiley-Interscience. p. 87. doi:10.1002/9780470316801. ISBN 9780471878766.

Starczewski, Artur, and Adam Krzyżak. "Performance Evaluation of the Silhouette Index."
International Conference on Artificial Intelligence and Soft Computing. Springer, Cham, 2015.
}
\author{
Igor D.S. Siciliani, Paulo H. dos Santos
}
\keyword{acca}
\keyword{silhouette,}
