% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_acca.R
\name{best_acca}
\alias{best_acca}
\alias{best_acca.cmatrix}
\alias{best_acca.matrix}
\title{Silhouette (clustering)}
\usage{
best_acca(m, ...)

\method{best_acca}{cmatrix}(m, mink, maxk, maxrep = 2L, maxiter = 100L, ...)

\method{best_acca}{matrix}(m, mink, maxk, maxrep = 2L, maxiter = 100L, ...)
}
\arguments{
\item{m}{\[\code{matrix(1)}]\cr correlation matrix
from \code{\link{corr_matrix}}.}

\item{...}{Additional arguments.}

\item{mink}{\[\code{integer(1)}]\cr minimum number of clusters considered.}

\item{maxk}{\[\code{integer(1)}]\cr maximum number of clusters considered.}

\item{maxrep}{\[\code{integer(1)}]\cr maximum number of interactions
without change in the clusters in the ACCA method.}

\item{maxiter}{\[\code{integer(1)}]\cr maximum number
of interactions in the ACCA method.}
}
\value{
\[\code{list(3)}]\cr A list with:
silhouette average with per k `$silhouette.ave`;
the sequence of clusters tested `$k` and
the optimal number of clusters `$best.k`.
}
\description{
Determining the optimal number of
cluster in the ACCA clustering using the
average silhouette aproach.
}
\examples{

x <- corrp::corrp(iris)
m <- corrp::corr_matrix(x)
best_acca(m, 2, 6)

}
\references{
Leonard Kaufman; Peter J. Rousseeuw (1990).
Finding groups in data : An introduction to cluster analysis.
Hoboken, NJ: Wiley-Interscience.
p. 87. doi:10.1002/9780470316801. ISBN 9780471878766.

Starczewski, Artur, and Adam Krzyżak.
"Performance evaluation of the silhouette index.
"International Conference on Artificial Intelligence
and Soft Computing. Springer, Cham, 2015.
}
\seealso{
\code{\link{sil_acca}}
}
\author{
Igor D.S. Siciliani
}
\keyword{,}
\keyword{acca}
\keyword{k}
\keyword{optimal}
\keyword{silhouette}
