% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\alias{corr_matrix.default}
\alias{corr_matrix.clist}
\title{Create Correlation Matrix from corrp Inferences}
\usage{
corr_matrix(c, ...)

\method{corr_matrix}{default}(c, col = c("infer.value", "stat.value", "isig"), isig = TRUE, ...)

\method{corr_matrix}{clist}(c, col = c("infer.value", "stat.value", "isig"), isig = TRUE, ...)
}
\arguments{
\item{c}{[\code{clist(1)}]\cr Output from the \code{\link{corrp}} function.}

\item{...}{Not used. Included for S3 method consistency.}

\item{col}{[\code{character(1)}]\cr Specifies the column to be used in the correlation matrix.}

\item{isig}{[\code{logical(1)}]\cr Determines whether values that are not statistically significant
should be represented by NA or FALSE in the correlation matrix.}
}
\description{
Using the results obtained from the corrp function,
this function creates a correlation matrix.
}
\examples{

iris_cor <- corrp(iris)
iris_m <- corr_matrix(iris_cor, isig = FALSE)
if (require("corrplot")) {
  corrplot(iris_m) # You can visualize the matrix using corrplot
}
}
\author{
Igor D.S. Siciliani, Paulo H. dos Santos
}
\keyword{correlation}
\keyword{corrp}
\keyword{matrix,}
