% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_rm.R
\name{corr_rm}
\alias{corr_rm}
\alias{corr_rm.clist}
\alias{corr_rm.list}
\alias{corr_rm.cmatrix}
\alias{corr_rm.matrix}
\title{corr_rm filter data.frame by correlation values.}
\usage{
corr_rm(df, c, ...)

\method{corr_rm}{clist}(
  df,
  c,
  col = c("infer.value", "stat.value"),
  isig = TRUE,
  cutoff = 0.75,
  ...
)

\method{corr_rm}{list}(
  df,
  c,
  col = c("infer.value", "stat.value"),
  isig = TRUE,
  cutoff = 0.75,
  ...
)

\method{corr_rm}{cmatrix}(df, c, cutoff = 0.75, ...)

\method{corr_rm}{matrix}(df, c, cutoff = 0.75, ...)
}
\arguments{
\item{df}{\[\code{data.frame(1)}]\cr input data frame.}

\item{c}{\[\code{clist(1)} | \code{cmatrix(1)}]\cr correlation list  output from \code{\link{corrp}} or
correlation matrix output from \code{\link{corr_matrix}}.}

\item{...}{Additional arguments (TODO).}

\item{col}{\[\code{character(1)}]\cr choose the column to be used in the correlation matrix}

\item{isig}{\[\code{logical(1)}]\cr values that are not statistically significant will
be represented by NA or FALSE in the correlation matrix.}

\item{cutoff}{\[\code{numeric(1)}]\cr A numeric value for the pair-wise absolute correlation cutoff.
The default values is 0.75.}
}
\description{
Remove highly correlated variables from a data.frame using the
corrp functions outputs and the caret package function \code{\link[caret]{findCorrelation}}.
}
\author{
Igor D.S. Siciliani
}
\keyword{,}
\keyword{clist}
\keyword{cmatrix}
\keyword{correlated}
\keyword{highly}
