% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrp2.R
\name{corrP}
\alias{corrP}
\title{corrp compute correlations types analysis in parallel backend.}
\usage{
corrP(df, ...)
}
\arguments{
\item{df}{input data frame.}

\item{...}{Additional arguments (TODO).}
}
\description{
same function as \code{corrp}.
}
