% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrP.R
\name{corrP}
\alias{corrP}
\title{corrP}
\usage{
corrP(df, parallel = TRUE, n.cores = 1, p.value = 0.05)
}
\arguments{
\item{df}{input data frame}

\item{parallel}{if is TRUE run the operations in parallel backend.}

\item{n.cores}{The number of cores to use for parallel execution.}

\item{p.value}{p-value probability of obtaining the observed results of a test,
assuming that the null hypothesis is correct. By default p.value=0.05.}
}
\description{
Compute correlations of columns of a large dataframe of mixed types
with parallel backend.
  The dataframe is allowed to have columns of these four classes: integer,
  numeric, factor and character. The character column is considered as
  categorical variable. This Package is based on Srikanth KS (talegari) cor2 function.
}
\details{
The correlation is computed as follows: \itemize{

  \item integer/numeric pair: pearson correlation using `cor` function. The
  valuelies between -1 and 1.

  \item integer/numeric - factor/categorical pair: correlation coefficient or
  squared root of R^2 coefficient of linear regression of integer/numeric
  variable over factor/categorical variable using `lm` function. The value
  lies between 0 and 1.

  \item factor/categorical pair: cramersV value is
  computed based on chisq test using `lsr::cramersV` function. The value lies
  between 0 and 1.

   \item All statistical tests are controlled by the confidence internal of
   p.value param. If the statistical tests do not obtain a significance lower
   than p.value, by default the correlation between variables will be zero.

   \item If any errors occur during operations by default the correlation will be zero.
   }

  For a comprehensive implementation, use `polycor::hetcor`
}
\examples{
air_cor <- corrP(airquality)
corrplot::corrplot(air_cor)
corrgram::corrgram(air_cor)

}
\author{
IDS siciliani (igor-siciliani)
}
\keyword{(http://choosealicense.com/licenses/agpl-3.0/)}
\keyword{AGPLv3}
\keyword{GNU}
