% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrp.R
\name{corrp}
\alias{corrp}
\title{corrp compute correlations types analysis in parallel backend.}
\usage{
corrp(
  df,
  parallel = TRUE,
  n.cores = 1,
  p.value = 0.05,
  verbose = TRUE,
  ptest.n.sum = 1000,
  ptest.r = F,
  comp = c("greater", "less"),
  alternative = c("two.sided", "less", "greater"),
  cor.nn = c("pearson", "mic", "dcor", "pps"),
  cor.nc = c("lm", "pps"),
  cor.cc = c("cramersV", "uncoef", "pps"),
  lm.args = list(),
  pearson.args = list(),
  dcor.args = list(),
  mic.args = list(),
  pps.args = list(),
  cramersV.args = list(),
  uncoef.args = list(),
  ...
)
}
\arguments{
\item{df}{input data frame.}

\item{parallel}{\[\code{logical(1)}\]\cr If its TRUE run the operations in parallel backend.}

\item{n.cores}{\[\code{numeric(1)}\]\cr The number of cores to use for parallel execution.}

\item{p.value}{\[\code{logical(1)}\]\cr
P-value probability of obtaining the observed results of a test,
assuming that the null hypothesis is correct. By default p.value=0.05 (Cutoff value for p-value.).}

\item{verbose}{\[\code{logical(1)}\]\cr Activate verbose mode.}

\item{comp}{\[\code{character(1)}\]\cr The param \code{p.value} must be greater
or less than those estimated in tests and correlations.}

\item{alternative}{\[\code{character(1)}\]\cr a character string specifying the alternative hypothesis for
the correlation inference. It must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{...}{Additional arguments (TODO).}
}
\description{
Compute correlations type analysis on mixed classes columns of larges dataframes
with parallel backend.
The dataframe is allowed to have columns of these four classes: integer,
numeric, factor and character. The character column is considered as
categorical variable. This method is original based on Srikanth KS (talegari) cor2 function.
}
\section{Details (Types)}{


- \code{integer/numeric pair} Pearson correlation using `cor` function. The
  value lies between -1 and 1.
- \code{integer/numeric - factor/categorical pair} correlation coefficient or
  squared root of R^2 coefficient of linear regression of integer/numeric
  variable over factor/categorical variable using \code{\link[stats]{lm}} function. The value
  lies between 0 and 1.
- \code{factor/categorical pair} Cramer's V value is
  computed based on chisq test and using \code{\link[lsrcramersV]{lm}} function. The value lies
  between 0 and 1.
}

\section{Details (Statistics)}{

- All statistical tests are controlled by the confidence internal of
  p.value param. If the statistical tests do not obtain a significance lower/upper
  than p.value, by default the value of variable `isig` will be "No".
- If any errors occur during operations by default the correlation will be `NA`.
}

\examples{
\dontrun{

air_cor = corrp(airquality)
air_m = corr_matrix(air_cor,isig = F)
corrplot::corrplot(air_m)

}

}
\references{
KS Srikanth,sidekicks,cor2, 2020.
URL \url{https://github.com/talegari/sidekicks/}.
Paul van der Laken, ppsr,2021.
URL \url{https://github.com/paulvanderlaken/ppsr}.
}
\author{
IDS siciliani (igor-siciliani)
}
\keyword{(http://choosealicense.com/licenses/agpl-3.0/)}
\keyword{AGPLv3}
\keyword{GNU}
