% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrP.R
\name{corrP}
\alias{corrP}
\title{corrP}
\usage{
corrP(df, parallel = TRUE, n.cores = 1, p.value = 0.05, verbose = TRUE)
}
\arguments{
\item{df}{input data frame}

\item{parallel}{if is TRUE run the operations in parallel backend.}

\item{n.cores}{The number of cores to use for parallel execution.}

\item{p.value}{p-value probability of obtaining the observed results of a test,
assuming that the null hypothesis is correct. By default p.value=0.05.}
}
\description{
Compute correlations of columns of a large dataframe of mixed types
with parallel backend.
  The dataframe is allowed to have columns of these four classes: integer,
  numeric, factor and character. The character column is considered as
  categorical variable. This Package is based on Srikanth KS (talegari) cor2 function.
}
\details{

}
\examples{
HairEyeColor_cor <- corrP(HairEyeColor)
corrplot::corrplot(HairEyeColor_cor)
corrgram::corrgram(HairEyeColor_cor)

}
\author{
IDS siciliani (igor-siciliani)
}
\keyword{(http://choosealicense.com/licenses/agpl-3.0/)}
\keyword{AGPLv3}
\keyword{GNU}
