% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrp.R
\name{corrp}
\alias{corrp}
\title{corrp compute correlations types analysis in parallel backend.}
\usage{
corrp(
  df,
  parallel = TRUE,
  n.cores = 1,
  p.value = 0.05,
  verbose = TRUE,
  num.s = 1000,
  rk = F,
  comp = c("greater", "less"),
  alternative = c("two.sided", "less", "greater"),
  cor.nn = c("pearson", "mic", "dcor", "pps"),
  cor.nc = c("lm", "pps"),
  cor.cc = c("cramersV", "uncoef", "pps"),
  lm.args = list(),
  pearson.args = list(),
  dcor.args = list(),
  mic.args = list(),
  pps.args = list(),
  cramersV.args = list(),
  uncoef.args = list(),
  ...
)
}
\arguments{
\item{df}{[\code{data.frame(1)}]\cr input data frame.}

\item{parallel}{[\code{logical(1)}]\cr If its TRUE run the operations in parallel backend.}

\item{n.cores}{[\code{numeric(1)}]\cr The number of cores to use for parallel execution.}

\item{p.value}{[\code{logical(1)}]\cr
P-value probability of obtaining the observed results of a test,
assuming that the null hypothesis is correct. By default p.value=0.05 (Cutoff value for p-value.).}

\item{verbose}{[\code{logical(1)}]\cr Activate verbose mode.}

\item{num.s}{[\code{numeric(1)}]\cr Used in permutation test. The number of samples with
replacement created with y numeric vector.}

\item{rk}{[\code{logical(1)}]\cr Used in permutation test.
if its TRUE transform x, y numeric vectors with samples ranks.}

\item{comp}{[\code{character(1)}]\cr The param \code{p.value} must be greater
or less than those estimated in tests and correlations.}

\item{alternative}{[\code{character(1)}]\cr a character string specifying the alternative hypothesis for
the correlation inference. It must be one of "two.sided" (default), "greater" or "less".
You can specify just the initial letter.}

\item{cor.nn}{[\code{character(1)}]\cr
Choose correlation type to be used in integer/numeric pair inference.
The options are `pearson: Pearson Correlation`,`mic: Maximal Information Coefficient`,
`dcor: Distance Correlation`,`pps: Predictive Power Score`.Default is `Pearson Correlation`.}

\item{cor.nc}{[\code{character(1)}]\cr
Choose correlation type to be used in integer/numeric - factor/categorical pair inference.
The option are `lm: Linear Model`,`pps: Predictive Power Score`. Default is `Linear Model`.}

\item{cor.cc}{[\code{character(1)}]\cr
Choose correlation type to be used in factor/categorical pair inference.
The option are `cramersV: Cramer's V`,`uncoef: Uncertainty coefficient`,
`pps: Predictive Power Score`. Default is ` Cramer's V`.}

\item{lm.args}{[\code{list(1)}]\cr additional parameters for the specific method.}

\item{pearson.args}{[\code{list(1)}]\cr additional parameters for the specific method.}

\item{dcor.args}{[\code{list(1)}]\cr additional parameters for the specific method.}

\item{mic.args}{[\code{list(1)}]\cr additional parameters for the specific method.}

\item{pps.args}{[\code{list(1)}]\cr additional parameters for the specific method.}

\item{cramersV.args}{[\code{list(1)}]\cr additional parameters for the specific method.}

\item{uncoef.args}{[\code{list(1)}]\cr additional parameters for the specific method.}

\item{...}{Additional arguments (TODO).}
}
\description{
Compute correlations analysis on mixed classes columns of larges dataframes
with parallel backend.
The dataframe is allowed to have columns of these four classes: integer,
numeric, factor and character. The character column is considered as
categorical variable.
}
\section{Details (Types)}{


- \code{integer/numeric pair} Pearson Correlation using \code{\link[stats]{cor}} function. The
  value lies between -1 and 1.\cr
- \code{integer/numeric pair} Distance Correlation using \code{\link[energy]{dcorT.test}} function. The
  value lies between 0 and 1.\cr
- \code{integer/numeric pair} Maximal Information Coefficient using \code{\link[minerva]{mine}} function. The
  value lies between 0 and 1.\cr
- \code{integer/numeric pair} Predictive Power Score using \code{\link[ppsr]{score}} function. The
  value lies between 0 and 1.\cr\cr
- \code{integer/numeric - factor/categorical pair} correlation coefficient or
  squared root of R^2 coefficient of linear regression of integer/numeric
  variable over factor/categorical variable using \code{\link[stats]{lm}} function. The value
  lies between 0 and 1.\cr
- \code{integer/numeric - factor/categorical pair} Predictive Power Score using \code{\link[ppsr]{score}} function. The
  value lies between 0 and 1.\cr\cr
- \code{factor/categorical pair} Cramer's V value is
  computed based on chisq test and using \code{\link[lsr]{cramersV}} function. The value lies
  between 0 and 1.\cr
- \code{factor/categorical pair} Uncertainty coefficient using \code{\link[DescTools]{UncertCoef}} function. The
  value lies between 0 and 1.\cr
- \code{factor/categorical pair} Predictive Power Score using \code{\link[ppsr]{score}} function. The
  value lies between 0 and 1.\cr
}

\section{Details (Statistics)}{

- All statistical tests are controlled by the confidence internal of
  p.value param. If the statistical tests do not obtain a significance greater/less
  than p.value, by default the value of variable `isig` will be `FALSE`.\cr
- There is no statistical significance test for the pps algorithm. By default isig = TRUE.\cr
- If any errors occur during operations by default the correlation will be `NA`.
}

\examples{
\dontrun{

air_cor = corrp(airquality)
air_m = corr_matrix(air_cor,isig = F)
corrplot::corrplot(air_m)

}

}
\references{
KS Srikanth,sidekicks,cor2, 2020.
URL \url{https://github.com/talegari/sidekicks/}.


Paul van der Laken, ppsr,2021.
URL \url{https://github.com/paulvanderlaken/ppsr}.
}
\author{
Igor D.S. Siciliani
}
\keyword{,}
\keyword{V}
\keyword{biserial}
\keyword{correlation}
\keyword{cramer's}
\keyword{distance}
\keyword{linear}
\keyword{mic}
\keyword{model}
\keyword{pearson}
\keyword{point}
\keyword{power}
\keyword{predictive}
\keyword{score}
