% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{ptest}
\alias{ptest}
\title{Correlation Permutation Test}
\usage{
ptest(
  x,
  y,
  FUN,
  rk = FALSE,
  alternative = c("greater", "less", "two.sided"),
  num.s = 250,
  ...
)
}
\arguments{
\item{x}{[\code{numeric(1)}]\cr A numeric vector.}

\item{y}{[\code{numeric(1)}]\cr A numeric vector.}

\item{FUN}{[\code{function(1)}]\cr The function to be applied.}

\item{rk}{[\code{logical(1)}]\cr If TRUE, transform x and y numeric vectors with sample ranks.}

\item{alternative}{[\code{character(1)}]\cr A character string specifying the alternative hypothesis.
Must be one of "greater" (default), "less", or "two.sided". You can specify just the initial letter.}

\item{num.s}{[\code{numeric(1)}]\cr The number of samples with replacement created from the y numeric vector.}

\item{...}{Additional arguments.}
}
\description{
Execute a one-sample permutation test on two numeric vectors.
One vector is kept constant while the other is "shuffled" by resampling.
This approximates the null hypothesis — that there is no dependency or difference between the variables.
}
\examples{

x <- iris[[1]]
y <- iris[[2]]
ptest(x, y, FUN = function(x, y) cor(x, y), alternative = "t")

}
