% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{ptest}
\alias{ptest}
\title{Correlation permutation test}
\usage{
ptest(
  x,
  y,
  FUN,
  r = F,
  alternative = c("two.sided", "less", "greater"),
  num.s = 1000,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{r}{logical. If its TRUE transform x, y numeric vectors with samples ranks.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{num.s}{number of samples with replacement created with y numeric vector.}
}
\description{
Execute one-sample permutation test on two numeric vector.
Its keep one vector constant and ‘shuffle’ the other by resampling.
This approximates the null hypothesis — that there is no dependency/difference between the variables.
}
\examples{
\dontrun{
x = iris[[1]]; y = iris[[2]]


ptest(x,y,FUN = function(x,y) cor(x,y) ,alternative = 't')


}
}
