% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{ptest}
\alias{ptest}
\title{Correlation permutation test}
\usage{
ptest(
  x,
  y,
  FUN,
  rk = FALSE,
  alternative = c("greater", "less", "two.sided"),
  num.s = 250,
  ...
)
}
\arguments{
\item{x}{[\code{numeric(1)}]\cr a numeric vector.}

\item{y}{[\code{numeric(1)}]\cr a numeric vector.}

\item{FUN}{[\code{function(1)}]\cr the function to be applied}

\item{rk}{[\code{logical(1)}]\cr if its TRUE transform x, y numeric vectors with samples ranks.}

\item{alternative}{[\code{character(1)}]\cr a character string specifying the alternative hypothesis,
must be one of "greater" (default), "less" or "two.sided". You can specify just the initial letter.}

\item{num.s}{[\code{numeric(1)}]\cr number of samples with replacement created with y numeric vector.}

\item{...}{Additional arguments.}
}
\description{
Execute one-sample permutation test on two numeric vector.
Its keep one vector constant and ‘shuffle’ the other by resampling.
This approximates the null hypothesis — that there is no dependency/difference between the variables.
}
\examples{

x <- iris[[1]]
y <- iris[[2]]
ptest(x, y, FUN = function(x, y) cor(x, y), alternative = "t")

}
