% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{ptest}
\alias{ptest}
\title{Correlation permutation test}
\usage{
ptest(
  x,
  y,
  FUN,
  rk = F,
  alternative = c("two.sided", "less", "greater"),
  num.s = 1000,
  ...
)
}
\arguments{
\item{x}{\[\code{numeric(1)}\]\cr a numeric vector.}

\item{y}{\[\code{numeric(1)}\]\cr a numeric vector.}

\item{rk}{\[\code{logical(1)}\]\cr if its TRUE transform x, y numeric vectors with samples ranks.}

\item{alternative}{\[\code{character(1)}\]\cr a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{num.s}{\[\code{numeric(1)}\]\cr number of samples with replacement created with y numeric vector.}
}
\description{
Execute one-sample permutation test on two numeric vector.
Its keep one vector constant and ‘shuffle’ the other by resampling.
This approximates the null hypothesis — that there is no dependency/difference between the variables.
}
\examples{
\dontrun{

x = iris[[1]]; y = iris[[2]]
ptest(x,y,FUN = function(x,y) cor(x,y) ,alternative = 't')

}
}
