% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{sil_acca}
\alias{sil_acca}
\alias{sil_acca.acca_list}
\alias{sil_acca.list}
\title{Silhouette (clustering)}
\usage{
sil_acca(acca, ...)

\method{sil_acca}{acca_list}(acca, m)

\method{sil_acca}{list}(acca, m)
}
\arguments{
\item{m}{[\code{matrix(1)}]\cr correlation matrix from \code{\link{corr_matrix}}}

\item{l}{[\code{acca_list(1)}]\cr Acca clustering results from \code{\link{acca}}}
}
\value{
[\code{numeric(1)}]\cr the silhouette coefficient for the maximum value of
 the mean {\displaystyle s(i)}s(i) over all data of the entire dataset
}
\description{
A C++ implementation of the Silhouette method
of interpretation and validation of consistency within acca clusters of data.
}
\examples{
\dontrun{

x = corrp::corrp(iris)
m = corrp::corr_matrix(x)
acca = corrp::acca(m,2)
sil_acca(acca,m)

}

}
\references{
Leonard Kaufman; Peter J. Rousseeuw (1990).
Finding groups in data : An introduction to cluster analysis.
Hoboken, NJ: Wiley-Interscience. p. 87. doi:10.1002/9780470316801. ISBN 9780471878766.

Starczewski, Artur, and Adam Krzyżak. "Performance evaluation of the silhouette index.
" International Conference on Artificial Intelligence and Soft Computing. Springer, Cham, 2015.
}
\author{
Igor D.S. Siciliani
}
\keyword{,}
\keyword{acca}
\keyword{silhouette}
